SIMSBIOHUB-445: Continued refactor
