+++ app/src/features/surveys/view/components/spatial-data/SurveySpatialObservationDataTable.tsx
