+++ app/src/features/surveys/view/components/spatial-data/SurveySpatialToolbar.tsx
